fix xades and pades on both jobs, still no valid pdf from pades on either
